Avoid type duplication
